code/langgraph-graph-rag/pyproject.toml
[tool.poetry]
name = "langgraph-graph-rag"
version = "0.1.0"
description = "A template for a Graph RAG LangGraph app deployable on IBM Cloud as an ai_service."
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "langgraph_graph_rag", from = "src" }
]


[tool.poetry.dependencies]
python = ">=3.11,<3.13"
langchain-ibm = "^0.3.9"
ibm-watsonx-ai = { version = ">=1.3.5", python = ">=3.11,<3.13" }
langgraph = ">0.2,<0.3"
langchain = "^0.3.21"
langchain_neo4j = "^0.4.0"
ibm_secrets_manager_sdk = "^2.1.9"
neo4j-graphrag = "1.7.0"  # due to the fact that rt24.1 is compiled with numpy<2
openai = "^1.101.0"  # needed because of bug in neo4j-graphrag
traceloop-sdk = "^0.47.3"
langfuse = "3.3.3"
docling = "^2.54.0"

[tool.poetry.group.dev.dependencies]
langchain-experimental = "^0.3.4"
langchain-text-splitters = "^0.3.9"
python-dotenv = "^1.0.1"
ibm-watsonx-ai-cli = ">=0.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
